requirements.txt
streamlit==1.56.0
google-genai==1.73.1
langgraph==1.1.8
pdfplumber==0.11.9
python-dotenv==1.2.2
pywebview==6.2.1
pythonnet==3.0.5
edge-tts==6.1.12
aiohttp==3.11.11
